Upgrade Branch dependency to 2.5.7
Kit must now have a minimum SDK version of 15, Since branch is no longer supporting < 4.0.3 Branch now relies on Application lifecycle callbacks. Because of this, there is no need to implement the KitIntegration.ActivityListener, but also it is imperative that we pass in the Application context, rather than any context in the call to Branch.getAutoInstance() in onKitCreate(), otherwise Branch will not start tracking sessions.